adds usage examples and tests, and linting
